+++ demo/generators/index.demo.php
@@ -5,4 +5,6 @@ require_once __DIR__ . '/../../vendor/autoload.php';
 use Katheroine\Layin\Demo\IndexPreconfiguredPageRenderer;
 
 $pageRenderer = new IndexPreconfiguredPageRenderer();
-$pageRenderer->renderPreconfiguredPage('index.layin.twig.html');
+$pageRenderer->setTemplateName('index.layin');
+
+echo $pageRenderer->render();