+++ site/public/pages/layouts/board/one_column.layin.php
@@ -5,4 +5,6 @@ require_once __DIR__ . '/../../../../../vendor/autoload.php';
 use Katheroine\Layin\Preconfiguration\LayoutsBoardPreconfiguredPageRenderer;
 
 $pageRenderer = new LayoutsBoardPreconfiguredPageRenderer();
-$pageRenderer->renderPreconfiguredPage('one_column.layin.twig.html');
+$pageRenderer->setTemplateName('one_column.layin');
+
+echo $pageRenderer->render();