+++ tests/unit_tests/renderers/AbstractPageRendererTest.php
@@ -33,70 +33,103 @@ class AbstractPageRendererTest extends TestCase
         );
     }
 
-    public function testSetTemplatesDirAbsolutePathFunctionExists()
+    public function testSetTemplatesDirPathFunctionExists()
     {
         $this->assertTrue(
             method_exists(
                 'Katheroine\Layin\Renderer\AbstractPageRenderer',
-                'setTemplatesDirAbsolutePath'
+                'setTemplatesDirPath'
             )
         );
     }
 
-    public function testSetTemplatesDirAbsolutePathReturnsSelf()
+    public function testSetTemplatesDirPathReturnsSelf()
     {
         $pageRenderer = new ConcretePageRenderer();
 
-        $result = $pageRenderer->setTemplatesDirAbsolutePath(__DIR__ . '/../../testing_environment/templates');
+        $result = $pageRenderer->setTemplatesDirPath(__DIR__ . '/../../testing_environment/templates');
 
-        $this->assertInstanceOf(ConcretePageRenderer::class, $result);
+        $this->assertSame($pageRenderer, $result);
     }
 
-    public function testSetTemplatesDirAbsolutePathWhenTemplateDirAbsolutePathIsNotString()
+    public function testSetTemplatesDirPathWhenTemplateDirAbsolutePathIsNotString()
     {
-        $templateDirAbsolutePath = 1024;
+        $templatesDirPath = 1024;
         $pageRenderer = new ConcretePageRenderer();
 
         $expectedErrorMessagePattern =
-            '/AbstractPageRenderer\:\:setTemplatesDirAbsolutePath\(\)\: '
-            . 'Argument \#1 \(\$templatesDirAbsolutePath\) must be of type string, int given/';
+            '/AbstractPageRenderer\:\:setTemplatesDirPath\(\)\: '
+            . 'Argument \#1 \(\$templatesDirPath\) must be of type string, int given/';
         $this->expectError(\TypeError::class);
         $this->expectErrorMessageMatches($expectedErrorMessagePattern);
 
-        $pageRenderer->setTemplatesDirAbsolutePath($templateDirAbsolutePath);
+        $pageRenderer->setTemplatesDirPath($templatesDirPath);
     }
 
-    public function testSetTemplateLocalPathFunctionExists()
+    public function testSetTemplateSubdirPathFunctionExists()
     {
         $this->assertTrue(
             method_exists(
                 'Katheroine\Layin\Renderer\AbstractPageRenderer',
-                'SetTemplateLocalPath'
+                'setTemplateSubdirPath'
             )
         );
     }
 
-    public function testSetTemplateLocalPathReturnsSelf()
+    public function testSetTemplateSubdirPathReturnsSelf()
     {
         $pageRenderer = new ConcretePageRenderer();
 
-        $result = $pageRenderer->SetTemplateLocalPath('subpages/');
+        $result = $pageRenderer->setTemplateSubdirPath('subpages/');
 
-        $this->assertInstanceOf(ConcretePageRenderer::class, $result);
+        $this->assertSame($pageRenderer, $result);
     }
 
-    public function testSetTemplateLocalPathWhenTemplateLocalPathIsNotString()
+    public function testSetTemplateSubdirPathWhenTemplateSubdirPathIsNotString()
     {
-        $templateLocalPath = 1024;
+        $templateSubdirPath = 1024;
         $pageRenderer = new ConcretePageRenderer();
 
         $expectedErrorMessagePattern =
-            '/AbstractPageRenderer\:\:setTemplateLocalPath\(\)\: '
-            . 'Argument \#1 \(\$templateLocalPath\) must be of type string, int given/';
+            '/AbstractPageRenderer\:\:setTemplateSubdirPath\(\)\: '
+            . 'Argument \#1 \(\$templateSubdirPath\) must be of type string, int given/';
         $this->expectError(\TypeError::class);
         $this->expectErrorMessageMatches($expectedErrorMessagePattern);
 
-        $pageRenderer->SetTemplateLocalPath($templateLocalPath);
+        $pageRenderer->setTemplateSubdirPath($templateSubdirPath);
+    }
+
+    public function testSetTemplateFileExtensionFunctionExists()
+    {
+        $this->assertTrue(
+            method_exists(
+                'Katheroine\Layin\Renderer\AbstractPageRenderer',
+                'setTemplateFileExtension'
+            )
+        );
+    }
+
+    public function testSetTemplateFileExtensionReturnsSelf()
+    {
+        $pageRenderer = new ConcretePageRenderer();
+
+        $result = $pageRenderer->setTemplateFileExtension('.html');
+
+        $this->assertSame($pageRenderer, $result);
+    }
+
+    public function testSetTemplateFileExtensionWhenFileExtensionIsNotString()
+    {
+        $templateFileExtension = 1024;
+        $pageRenderer = new ConcretePageRenderer();
+
+        $expectedErrorMessagePattern =
+            '/AbstractPageRenderer\:\:setTemplateFileExtension\(\)\: '
+            . 'Argument \#1 \(\$templateFileExtension\) must be of type string, int given/';
+        $this->expectError(\TypeError::class);
+        $this->expectErrorMessageMatches($expectedErrorMessagePattern);
+
+        $pageRenderer->setTemplateFileExtension($templateFileExtension);
     }
 
     public function testSetTemplateNameFunctionExists()
@@ -115,7 +148,7 @@ class AbstractPageRendererTest extends TestCase
 
         $result = $pageRenderer->SetTemplateName('page.twig.html');
 
-        $this->assertInstanceOf(ConcretePageRenderer::class, $result);
+        $this->assertSame($pageRenderer, $result);
     }
 
     public function testSetTemplateNameWhenTemplateNameIsNotString()
@@ -132,6 +165,39 @@ class AbstractPageRendererTest extends TestCase
         $pageRenderer->SetTemplateName($templateName);
     }
 
+    public function testSetTemplateParamsFunctionExists()
+    {
+        $this->assertTrue(
+            method_exists(
+                'Katheroine\Layin\Renderer\AbstractPageRenderer',
+                'setTemplateParams'
+            )
+        );
+    }
+
+    public function testSetTemplateParamsReturnsSelf()
+    {
+        $pageRenderer = new ConcretePageRenderer();
+
+        $result = $pageRenderer->SetTemplateParams([]);
+
+        $this->assertSame($pageRenderer, $result);
+    }
+
+    public function testSetTemplateParamsWhenTemplateNameIsNotString()
+    {
+        $templateParams = 1024;
+        $pageRenderer = new ConcretePageRenderer();
+
+        $expectedErrorMessagePattern =
+            '/AbstractPageRenderer\:\:setTemplateParams\(\)\: '
+            . 'Argument \#1 \(\$templateParams\) must be of type array, int given/';
+        $this->expectError(\TypeError::class);
+        $this->expectErrorMessageMatches($expectedErrorMessagePattern);
+
+        $pageRenderer->setTemplateParams($templateParams);
+    }
+
     public function testRenderFunctionExists()
     {
         $this->assertTrue(
@@ -142,101 +208,16 @@ class AbstractPageRendererTest extends TestCase
         );
     }
 
-    public function testRenderReturnsNothing()
+    public function testRenderReturnsString()
     {
         $pageRenderer = new ConcretePageRenderer();
         $pageRenderer
-            ->setTemplatesDirAbsolutePath(__DIR__ . '/../../testing_environment/templates')
-            ->SetTemplateName('page.twig.html');
+            ->setTemplatesDirPath(__DIR__ . '/../../testing_environment/templates')
+            ->setTemplateName('page')
+            ->setTemplateFileExtension('.twig.html');
 
-        ob_start(); // Doesn't allow to echo rendered template.
         $result = $pageRenderer->render();
-        ob_end_clean();
 
-        $this->assertNull($result);
-    }
-
-    public function testRenderWhenNoTemplateIsSet()
-    {
-        $pageRenderer = new ConcretePageRenderer();
-
-        // Twig doc: When the template cannot be found
-        $this->expectException(\Twig\Error\LoaderError::class);
-        $this->expectExceptionMessage('There are no registered paths for namespace "__main__".');
-
-        $pageRenderer->render();
-    }
-
-    public function testRenderWhenTemplateContentIsSyntacticallyImproper()
-    {
-        $pageRenderer = new ConcretePageRenderer();
-        $pageRenderer
-            ->setTemplatesDirAbsolutePath(__DIR__ . '/../../testing_environment/templates')
-            ->SetTemplateName('syntactically_improper_content.twig.html');
-
-        // Twig doc: When an error occurred during compilation.
-        $this->expectException(\Twig\Error\SyntaxError::class);
-        $this->expectExceptionMessage('Unexpected "}"');
-
-        $pageRenderer->render();
-    }
-
-    public function testRenderWithoutTemplateLocalPath()
-    {
-        $pageRenderer = new ConcretePageRenderer();
-        $pageRenderer
-            ->setTemplatesDirAbsolutePath(__DIR__ . '/../../testing_environment/templates')
-            ->SetTemplateName('page.twig.html');
-
-        ob_start(); // Allow to capture rendered content instead of echoing it.
-        $pageRenderer->render();
-        $actualRenderedContent = ob_get_contents();
-        ob_end_clean();
-
-        $expectedRenderedContent = "<!doctype html>\n"
-            . "<html lang=\"en-GB\">\n"
-            . "  <head>\n"
-            . "    <meta charset=\"\">\n"
-            . "    <meta name=\"language\" content=\"english\">\n"
-            . "    <meta name=\"description\" content=\"All purpose web page layout\">\n"
-            . "    <meta name=\"keywords\" content=\"layout, web page\">\n"
-            . "    <meta name=\"author\" content=\"usagi <usagi@moon.com>\">\n"
-            . "  </head>\n"
-            . "  <body>\n"
-            . "    <h1>Welcome on the home page!</h1>\n"
-            . "  </body>\n"
-            . "</html>\n";
-
-        $this->assertEquals($expectedRenderedContent, $actualRenderedContent);
-    }
-
-    public function testRenderWithTemplateLocalPath()
-    {
-        $pageRenderer = new ConcretePageRenderer();
-        $pageRenderer
-            ->setTemplatesDirAbsolutePath(__DIR__ . '/../../testing_environment/templates')
-            ->setTemplateLocalPath('subpages/')
-            ->SetTemplateName('subpage.twig.html');
-
-        ob_start(); // Allow to capture rendered content instead of echoing it.
-        $pageRenderer->render();
-        $actualRenderedContent = ob_get_contents();
-        ob_end_clean();
-
-        $expectedRenderedContent = "<!doctype html>\n"
-            . "<html lang=\"en-GB\">\n"
-            . "  <head>\n"
-            . "    <meta charset=\"\">\n"
-            . "    <meta name=\"language\" content=\"english\">\n"
-            . "    <meta name=\"description\" content=\"All purpose web page layout\">\n"
-            . "    <meta name=\"keywords\" content=\"layout, web page\">\n"
-            . "    <meta name=\"author\" content=\"usagi <usagi@moon.com>\">\n"
-            . "  </head>\n"
-            . "  <body>\n"
-            . "    <h1>Welcome on the subpage!</h1>\n"
-            . "  </body>\n"
-            . "</html>\n";
-
-        $this->assertEquals($expectedRenderedContent, $actualRenderedContent);
+        $this->assertIsString($result);
     }
 }